Primeira vaga: 67 declarações fora do ponto flutuante, e a área do glifo em inteiros.
O Aarão: «eliminar os doubles é prioridade», e o plano é float → racional em long → int →
uint8, porque a reta real já está construída em binário no geometrico e o flutuante é
INCOMPATÍVEL com o que lá está provado. Até agora eu vinha a acrescentar caminhos exactos
AO LADO dos doubles, e o levantamento não descia. Agora o double SAI.

    declarados   2878 -> 2811
    a alimentar   717 -> 701

SPLINE §P3, e este é o achado. As coordenadas de um glifo são INTEIRAS — `Pt` tem
`long x, y` — logo a área é um racional exacto e os cinco doubles só a transportavam. Mas o
caminho até eles PERDIA: o ponto da curva implícito entre dois controlos é o ponto MÉDIO, e
`(a.x + b.x)/2` em inteiros trunca quando a soma é ímpar.

    1105 dos 1436 pontos médios implícitos da fonte do documento têm soma ÍMPAR — 77%

Dobram-se as coordenadas, o médio passa a exacto, e a área sai como 96·A em `long`:
M=43110164, O=34167368, o=22965032, i=10834560, espaço=0. As funções area2_quad,
area2_reta e area_glifo ficaram sem chamador e foram apagadas.

DIF §F3 e §F4: fora o bloco de dezasseis pares em diferenças finitas (h = 1e-5, limiar
1e-6) e o laço do Euler em pow(). As mesmas leis ficam medidas nos blocos exactos — F⁴ = id
e F² = paridade em ℤ₁₇ com resíduo zero, Leibniz e linearidade em 15625 pares de polinómios
inteiros. `double` no ficheiro: 87 -> 72.

E 36 FICHEIROS por varredura automática, cada troca aceite só com saída IDÊNTICA byte a byte.

O ERRO DE EXECUÇÃO, e vai dito porque custou uma bateria vermelha: lancei a varredura em
oito processos que NÃO ficaram divididos em fatias. Escreveram nos mesmos ficheiros e
invalidaram as verificações uns dos outros — `long ant = s` a truncar um double no
corpo_fisico, e `long h = 1e-8` a virar ZERO no cosmico, laço infinito e timeout. Dois
vermelhos. Revertidos.

E depois varri as 36 à procura do mesmo estrago por outra via: trocas cujo valor inicial não
é inteiro. Mais duas, que tinham passado por a saída não mudar — `long TOL = 1e-9` e
`long A_colheita = 100e-4`, ambas a virar zero. Revertidas.

Bateria 501:501.

+++ tests/analog.c
@@ -483,7 +483,7 @@ static void B5_soma(void) {
     long passou = 0, tot = 0; int arvore_torta_bateu = 0;
     for (int seed = 0; seed < 200; seed++) {
         double v[8]; int nb = 2 + seed % 7;                 /* de 2 a 8 fontes */
-        double linear = 0;                                  /* fontes > 0: o ramo */
+        long linear = 0;                                  /* fontes > 0: o ramo */
         for (int j = 0; j < nb; j++) { v[j] = 1 + (seed*7 + j*13) % 20; linear += v[j]; }
         if (soma_arvore(v, nb) == linear) passou++;             /* árvore == linear */
         /* o dente: o topo subtrai os ramos. Com fontes > 0 o ramo direito é > 0,
